langgraph/requirements.txt
python-dotenv
llama-index
llama-index-vector-stores-qdrant 
llama-index-readers-file 
llama-index-embeddings-fastembed 
llama-index-llms-openai
llama-index-llms-groq
qdrant_client
fastembed
langgraph
langchain
langchain-core
langchain-community
langchain-openai
typing
markitdown
crawl4ai
pyquery
playwright
trustcall
langgraph-cli
google
tavily-python
git+https://github.com/stanfordnlp/dspy.git